somfy_rts/requirements.txt

paho-mqtt~=1.4
pyyaml~=5.1
